backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==3.2.0",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "uvicorn>=0.42.0",
]
